+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/useCaseCreationForm.ts
@@ -1,25 +1,25 @@
 import { TWorkflowDefinition } from '@/domains/workflows/fetchers';
 import { useCreateWorkflowMutation } from '@/domains/workflows/hooks/mutations/useCreateWorkflowMutation/useCreateWorkflowMutation';
-import { prepareContext } from '@/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/utils/prepare-context';
+import { createContextFromFormData } from '@/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/utils/create-context-from-form-data';
 import { useCaseCreationContext } from '@/pages/Entities/components/CaseCreation/context/case-creation-context/hooks/useCaseCreationContext';
 import { AnyObject } from '@ballerine/ui';
 import { useCallback } from 'react';
 
 export const useCaseCreationForm = (workflowDefinition: TWorkflowDefinition) => {
   const { isLoading, mutateAsync } = useCreateWorkflowMutation();
-  const { isMultipleCasesCreation, setIsOpen: setOpen } = useCaseCreationContext();
+  const { isMultipleCasesCreation, setIsOpen } = useCaseCreationContext();
 
   const handleSubmit = useCallback(
     async (formData: AnyObject) => {
-      const context = prepareContext(formData);
+      const context = createContextFromFormData(formData);
 
       await mutateAsync({ workflowDefinitionId: workflowDefinition.id, context });
 
       if (!isMultipleCasesCreation) {
-        setOpen(false);
+        setIsOpen(false);
       }
     },
-    [workflowDefinition, isMultipleCasesCreation, mutateAsync, setOpen],
+    [workflowDefinition, isMultipleCasesCreation, mutateAsync, setIsOpen],
   );
 
   return {
