+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/utils/create-context-from-form-data.ts
@@ -2,7 +2,7 @@ import { TWorkflowById } from '@/domains/workflows/fetchers';
 import { generateEntityId } from '@/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseCreationForm/utils/generate-entity-id';
 import { AnyObject } from '@ballerine/ui';
 
-export const prepareContext = (formData: AnyObject): TWorkflowById['context'] => {
+export const createContextFromFormData = (formData: AnyObject): TWorkflowById['context'] => {
   const context = {
     entity: {
       id: generateEntityId(),
