+++ services/workflows-service/prisma/schema.prisma
@@ -150,9 +150,8 @@ model WorkflowDefinition {
   name            String
   version         Int     @default(1)
 
-  projectId               String?
-  isPublic                Boolean @default(false)
-  isManualCreationEnabled Boolean @default(false)
+  projectId String?
+  isPublic  Boolean @default(false)
 
   definitionType     String
   definition         Json
