+++ services/workflows-service/src/case-management/case-management.module.ts
@@ -1,7 +1,7 @@
 import { CaseManagementService } from '@/case-management/case-management.service';
 import { CaseManagementController } from '@/case-management/controllers/case-management.controller';
 import { WorkflowDefinitionModule } from '@/workflow-defintion/workflow-definition.module';
-import { WorkflowServiceModule } from '@/workflow/workflow-service.module';
+import { WorkflowServiceModule } from '@/workflow/workflow.module';
 import { Module } from '@nestjs/common';
 
 @Module({
