+++ services/workflows-service/src/case-management/types/transition-schema.ts
@@ -1,4 +1,4 @@
-export interface TransitionSchema {
+export interface ITransitionSchema {
   state: string;
   schema: Record<PropertyKey, unknown>;
   additionalParameters?: Record<PropertyKey, unknown>;
