+++ services/workflows-service/src/workflow/workflow-controller.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { WorkflowControllerExternal } from './workflow.controller.external';
+import { WorkflowControllerInternal } from './workflow.controller.internal';
+import { WorkflowServiceModule } from '@/workflow/workflow.module';
+
+@Module({
+  imports: [WorkflowServiceModule],
+  controllers: [WorkflowControllerExternal, WorkflowControllerInternal],
+  exports: [WorkflowServiceModule],
+})
+export class WorkflowControllerModule {}
